勤怠情報編集・削除完了

+++ SmartAttendance/src/main/java/com/example/demo/entity/RegistKintaiEntity.java
@@ -1,5 +1,6 @@
 package com.example.demo.entity;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 import lombok.Data;
@@ -12,8 +13,8 @@ public class RegistKintaiEntity {
 		private int start_M;
 		private int end_H;
 		private int end_M;
-		private double total_workingTime;/*実労働時間[時間]*/
-		private double breakTime; /*休憩時間[時間]*/
-		private double excessTime; /*超過労働時間[時間]*/
+		private BigDecimal total_workingTime;/*実労働時間[時間]*/
+		private int breakTime; /*休憩時間[時間]*/
+		private int excessTime; /*超過労働時間[時間]*/
 
 }
